requirements.txt
backcall==0.1.0
bleach==2.1.3
click==6.7
decorator==4.2.1
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-PageDown==0.2.2
Flask-WTF==0.14.2
html5lib==1.0.1
ipython==6.3.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.12
parso==0.1.1
pexpect==4.5.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycrypto==2.6.1
Pygments==2.2.0
PyMySQL==0.8.0
PyMySQL3==0.5
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
visitor==0.1.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
WTForms==2.1
